Settings: Move blur toggle to Display options
