+++ res/xml/development_settings.xml
@@ -577,10 +577,6 @@
             android:summary="@string/hwui_force_dark_summary" />
 
         <SwitchPreference
-            android:key="enable_blurs_on_windows"
-            android:title="@string/window_blurs"/>
-
-        <SwitchPreference
             android:key="force_msaa"
             android:title="@string/force_msaa"
             android:summary="@string/force_msaa_summary" />
